【Hackathon 5th No.105】move fusion_gru/fusion_seqconv_eltadd_relu/fusion_seqexpand_concat_fc to phi (#57881)
* add a part of fusion_gru

* add some code

* move the fusion_seqconv_eltadd_relu_op to phi but have the same bug in gru_op

* move fusion_seqexpand_concat_fc_op to phi, but have the same bug in fusion_gru_op

* fix the intermediate bug

* fix the conflict

* pass the fusion_seqconv in new ir

* try to move fusion_gru_mkldnn_kernel to phi

* move fusion_gru_mkldnn to phi, pass the test but some bug in new iR

* change some discribe

* fix conflict

* fix bug

* fix the bug in getInputName

* remove some describe

* add the set_dtype
