+++ paddle/fluid/operators/fused/CMakeLists.txt
@@ -13,7 +13,6 @@ register_operators(
   yolo_box_head_op
   yolo_box_post_op
   fusion_group_op
-  fusion_gru_op
   fusion_lstm_op
   fused_bn_add_activation_op
   fused_attention_op
@@ -27,8 +26,6 @@ register_operators(
   fused_gate_attention_op
   resnet_basic_block_op)
 
-# fusion_gru_op does not have CUDA kernel
-op_library(fusion_gru_op)
 op_library(fusion_lstm_op)
 if(WITH_AVX
    AND AVX512F_FOUND
