+++ paddle/phi/infermeta/fusion.cc
@@ -1918,8 +1918,8 @@ void FusedEmbeddingEltWiseLayerNormInferMeta(
 
   auto dim_output = phi::make_ddim({batch, seq_len, hidden});
   out->set_dims(dim_output);
-  // out->share_lod(ids);
-  // context->ShareLoD("Ids", /*->*/ "Out");
+  out->share_lod(*ids[0]);
+  out->set_dtype((*embs[0]).dtype());
 }
 
 void FusionTransposeFlattenConcatInferMeta(
@@ -1981,6 +1981,7 @@ void FusionTransposeFlattenConcatInferMeta(
     out_dims[concat_axis] = -1;
   }
   out->set_dims(phi::make_ddim(out_dims));
+  out->set_dtype((*x[0]).dtype());
 }
 
 void FusedFCElementwiseLayerNormInferMeta(const MetaTensor& x,
@@ -2162,13 +2163,304 @@ void FusedFCElementwiseLayerNormInferMeta(const MetaTensor& x,
   }
 
   out->set_dims(y_dims);
+  out->set_dtype(x.dtype());
   if (mean) {
+    mean->set_dtype(x.dtype());
     mean->set_dims({dim_0});
   }
   if (variance) {
     variance->set_dims({dim_0});
+    variance->set_dtype(x.dtype());
   }
   out->share_lod(x);
 }
 
+void FusionGRUInferMeta(const MetaTensor& x,
+                        const MetaTensor& h0,
+                        const MetaTensor& weight_x,
+                        const MetaTensor& weight_h,
+                        const MetaTensor& bias,
+                        const std::string& activation,
+                        const std::string& gate_activation,
+                        const bool is_reverse,
+                        const bool use_seq,
+                        const bool origin_mode,
+                        const bool use_mkldnn,
+                        const std::string& mkldnn_data_type,
+                        const float scale_data,
+                        const float shift_data,
+                        const std::vector<float>& scale_weights,
+                        const bool force_fp32_output,
+                        MetaTensor* reordered_h0,
+                        MetaTensor* xx,
+                        MetaTensor* batched_input,
+                        MetaTensor* batched_out,
+                        MetaTensor* hidden) {
+  std::string mkldnn_data_type_list[] = {"float32", "int8", "bfloat16"};
+  PADDLE_ENFORCE_EQ(
+      std::find(std::begin(mkldnn_data_type_list),
+                std::end(mkldnn_data_type_list),
+                mkldnn_data_type) != std::end(mkldnn_data_type_list),
+      true,
+      phi::errors::InvalidArgument("The mkldnn_data_type shoule be [float32, "
+                                   "int8, bfloat16], but found %s.",
+                                   mkldnn_data_type.c_str()));
+
+  DDim x_dims = x.dims();
+  auto x_mat_dims = (x_dims.size() == 3 && x_dims[1] == 1)
+                        ? phi::flatten_to_2d(x_dims, 1)
+                        : x_dims;
+  PADDLE_ENFORCE_EQ(
+      x_mat_dims.size(),
+      2,
+      phi::errors::InvalidArgument("The size of input X dims should be 2, "
+                                   "or 3 with second dimension equal to "
+                                   "1, but now Input X dim is:[%s] ",
+                                   x_dims));
+
+  auto wx_dims = weight_x.dims();
+  PADDLE_ENFORCE_EQ(wx_dims.size(),
+                    2,
+                    phi::errors::InvalidArgument(
+                        "The rank of Input(WeightX) should be 2, but received "
+                        "WeightX dim size is:%d, WeightX dim is:[%s] ",
+                        wx_dims.size(),
+                        wx_dims));
+  PADDLE_ENFORCE_EQ(
+      wx_dims[0],
+      x_mat_dims[1],
+      phi::errors::InvalidArgument(
+          "The first dimension of flattened WeightX"
+          "should equal to last dimension of flattened input X, but "
+          "received fattened WeightX dimension is:%d, flattened X dimension "
+          "is:%d",
+          wx_dims[0],
+          x_mat_dims[1]));
+
+  int frame_size = static_cast<int>(wx_dims[1] / 3);
+  auto wh_dims = weight_h.dims();
+
+  PADDLE_ENFORCE_EQ(wh_dims.size(),
+                    2,
+                    phi::errors::InvalidArgument(
+                        "The rank of Input(WeightH) should be 2, but received "
+                        "WeightH dim size is:%d, WeightH dim is:[%s]",
+                        wh_dims.size(),
+                        wh_dims));
+  PADDLE_ENFORCE_EQ(wh_dims[0],
+                    frame_size,
+                    phi::errors::InvalidArgument(
+                        "The first dimension of WeightH "
+                        "should equal to frame_size, but received WeightH's "
+                        "first dimension is: "
+                        "%d, frame size is:%d",
+                        wh_dims[0],
+                        frame_size));
+  PADDLE_ENFORCE_EQ(wh_dims[1],
+                    3 * frame_size,
+                    phi::errors::InvalidArgument(
+                        "The second dimension of Input(WeightH) "
+                        "should equal to 3 * frame_size, but received WeightH "
+                        "is:%d, frame size is:%d",
+                        wh_dims[1],
+                        frame_size));
+
+  if (h0) {
+    auto h0_dims = h0.dims();
+    PADDLE_ENFORCE_EQ(h0_dims[1],
+                      frame_size,
+                      phi::errors::InvalidArgument(
+                          "The width of H0 must be equal to frame_size, but "
+                          "receiced the width of H0 is:%d, frame size is:%d",
+                          h0_dims[1],
+                          frame_size));
+    reordered_h0->set_dtype(x.dtype());
+  }
+  if (bias) {
+    auto b_dims = bias.dims();
+    PADDLE_ENFORCE_EQ(b_dims.size(),
+                      2,
+                      phi::errors::InvalidArgument(
+                          "The rank of Input(Bias) should be 2, but received "
+                          "Bias rank is:%d, Bias dim is:[%s]",
+                          b_dims.size(),
+                          b_dims));
+    PADDLE_ENFORCE_EQ(b_dims[0],
+                      1,
+                      phi::errors::InvalidArgument(
+                          "The first dimension of Input(Bias) should be 1, but "
+                          "received Bias first dim is:%d, Bias dim is:[%s]",
+                          b_dims[0],
+                          b_dims));
+    PADDLE_ENFORCE_EQ(b_dims[1],
+                      frame_size * 3,
+                      phi::errors::InvalidArgument(
+                          "The shape of Bias must be [1, frame_size * 3], but "
+                          "received bias dim is:[%s], frame size is:%d",
+                          b_dims,
+                          frame_size));
+  }
+  DDim out_dims({x_mat_dims[0], frame_size});
+  hidden->set_dims(out_dims);
+  hidden->share_lod(x);
+  hidden->set_dtype(x.dtype());
+  int xx_width = 0;
+  if (use_seq) {
+    xx_width = static_cast<int>(wx_dims[1]);
+  } else {
+    xx_width = static_cast<int>(x_mat_dims[1] > wx_dims[1] ? wx_dims[1]
+                                                           : x_mat_dims[1]);
+    batched_input->set_dims({x_mat_dims[0], wx_dims[1]});
+    batched_input->set_dtype(x.dtype());
+    batched_out->set_dims(out_dims);
+    batched_out->set_dtype(x.dtype());
+  }
+  xx->set_dims({x_mat_dims[0], xx_width});
+  xx->set_dtype(x.dtype());
+  xx->share_lod(x);
+}
+
+void FusionSeqConvEltAddReluInferMeta(const MetaTensor& x,
+                                      const MetaTensor& filter,
+                                      const MetaTensor& bias,
+                                      const int context_length,
+                                      const int context_start,
+                                      const int context_stride,
+                                      MetaTensor* out,
+                                      MetaTensor* col_mat) {
+  auto x_dims = x.dims();
+  auto w_dims = filter.dims();
+  PADDLE_ENFORCE_GT(
+      context_length,
+      0,
+      phi::errors::InvalidArgument("context_length should be greater than 0, "
+                                   "but received context_length is: %d",
+                                   context_length));
+  PADDLE_ENFORCE_EQ(context_stride,
+                    1,
+                    phi::errors::InvalidArgument(
+                        "Currently, FusionSeqConvEltAddReluOp only supports "
+                        "contextStride=1, but received value is: %d.",
+                        context_stride));
+
+  PADDLE_ENFORCE_EQ(
+      x_dims.size(),
+      2,
+      phi::errors::InvalidArgument(
+          "Input(X) should be 2-D tensor, but reveiced value is: %d.",
+          x_dims.size()));
+
+  PADDLE_ENFORCE_EQ(
+      w_dims.size(),
+      2,
+      phi::errors::InvalidArgument(
+          "Filter should be 2-D tensor, but reveiced value is: %d.",
+          w_dims.size()));
+
+  PADDLE_ENFORCE_EQ(w_dims[0],
+                    context_length * x_dims[1],
+                    phi::errors::InvalidArgument(
+                        "Filter's height should be equal to context_length * "
+                        "input_hidden_size, but received Filter height is: %d,"
+                        "context_length is: %d, input_hidden_size is: %d.",
+                        w_dims[0],
+                        context_length,
+                        x_dims[1]));
+
+  PADDLE_ENFORCE_GT(
+      context_length + context_start,
+      0,
+      phi::errors::InvalidArgument(
+          "contextStart size should be smaller than contextLength, "
+          "but received context_length is: %d, contextStart is: "
+          "%d.",
+          context_length,
+          context_start));
+  out->set_dims({x_dims[0], w_dims[1]});
+  col_mat->set_dims({x_dims[0], w_dims[0]});
+  out->share_lod(x);
+  col_mat->set_dtype(x.dtype());
+  out->set_dtype(x.dtype());
+}
+
+void FusionSeqExpandConcatFCInferMeta(const std::vector<const MetaTensor*>& x,
+                                      const MetaTensor& fc_weight,
+                                      const MetaTensor& fc_bias,
+                                      const std::string& fc_activation,
+                                      MetaTensor* out,
+                                      MetaTensor* fc_out) {
+  PADDLE_ENFORCE_GT(x.size(),
+                    1UL,
+                    phi::errors::InvalidArgument(
+                        "Inputs(X) of FusionSeqExpandConcatFCOp should larger "
+                        "than 1, but received value is: %d.",
+                        x.size()));
+
+  std::vector<DDim> ins_dims;
+  ins_dims.reserve(x.size());
+  std::transform(x.begin(),
+                 x.end(),
+                 std::back_inserter(ins_dims),
+                 [](const MetaTensor* var) { return var->dims(); });
+
+  auto w_dims = fc_weight.dims();  // (M0+M1+M2+..) x D
+  PADDLE_ENFORCE_EQ(
+      w_dims.size(),
+      2,
+      phi::errors::InvalidArgument(
+          "Input(FCWeight)'s rank must be 2, but received value is: %d.",
+          w_dims.size()));
+  const int D = static_cast<int>(w_dims[1]);
+  int sum = static_cast<int>(ins_dims[0][1]);
+  for (size_t i = 1; i < ins_dims.size(); ++i) {
+    sum += static_cast<int>(ins_dims[i][1]);
+  }
+  PADDLE_ENFORCE_EQ(
+      sum,
+      w_dims[0],
+      phi::errors::InvalidArgument("FC height should be sum of all inputs "
+                                   "width, but received FC height is: %d, "
+                                   "sum of all inputs width is: %d.",
+                                   w_dims[0],
+                                   sum));
+  if (fc_bias) {
+    auto b_dims = fc_bias.dims();
+    PADDLE_ENFORCE_EQ(
+        b_dims.size() == 1 || b_dims.size() == 2,
+        true,
+        phi::errors::InvalidArgument(
+            "FCBias dim should be 1 or 2, but received value is: %d.",
+            b_dims.size()));
+    if (b_dims.size() == 1) {
+      PADDLE_ENFORCE_EQ(b_dims[0],
+                        D,
+                        phi::errors::InvalidArgument(
+                            "FCBias shapes must be %d when FCBias dim = 1, but "
+                            "received value is: %d.",
+                            D,
+                            b_dims[0]));
+    } else {
+      PADDLE_ENFORCE_EQ(b_dims[0],
+                        1,
+                        phi::errors::InvalidArgument(
+                            "FCBias shapes must be 1x%d, when FCBias dim = 2, "
+                            "but received dim[0] is: %d.",
+                            D,
+                            b_dims[0]));
+      PADDLE_ENFORCE_EQ(b_dims[1],
+                        D,
+                        phi::errors::InvalidArgument(
+                            "FCBias shapes must be 1x%d, when FCBias dim = 2, "
+                            "but received dim[1] is: %d.",
+                            D,
+                            b_dims[1]));
+    }
+  }
+  fc_out->set_dtype((*x[0]).dtype());
+  out->set_dims({ins_dims[0][0], D});
+  out->set_dtype((*x[0]).dtype());
+  // fcout should be reshape when run since can not get lod in infershape
+  // explicit share the ref lod
+  out->share_lod(*x[0]);
+}
 }  // namespace phi
