+++ test/legacy_test/test_fusion_seqconv_eltadd_relu_op.py
@@ -56,7 +56,7 @@ class TestSeqConvEltAddRelu(OpTest):
         self.outputs = {'Out': out}
 
     def test_check_output(self):
-        self.check_output()
+        self.check_output(check_dygraph=False)
 
 
 class TestSeqConvEltAddReluBS1(TestSeqConvEltAddRelu):
